+++ lib/timber/contexts/http_context.ex
@@ -1,10 +1,19 @@
 defmodule Timber.Contexts.HTTPContext do
-  @moduledoc """
-  The HTTP context tracks information about an HTTP request currently
-  being handled.
+  @moduledoc ~S"""
+  **DEPRECATED**
 
-  Note: Timber can automatically add context information about HTTP requests if
-  you use a `Plug` based framework through the `Timber.Plug.HTTPContext`.
+  This module is deprecated in favor of using `map`s. The next evolution of Timber (2.0)
+  no long requires a strict schema and therefore simplifies how users set context:
+
+      Timber.add_context(http: %{method: "post", path: "/checkout"})
+
+  In addition, you can use Timber integrations to automatically capture context in `Plug`
+  and `Phoenix`:
+
+  * [`:timber_phoenix`](https://github.com/timberio/timber-elixir-phoenix)
+  * [`:timber_plug`](https://github.com/timberio/timber-elixir-plug)
+
+  Checkout the `README` for a list of all integrations.
   """
 
   @type t :: %__MODULE__{
@@ -22,4 +31,11 @@ defmodule Timber.Contexts.HTTPContext do
         }
 
   defstruct [:method, :path, :request_id, :remote_addr]
+
+  defimpl Timber.Contextable do
+    def to_context(context) do
+      context = Map.from_struct(context)
+      %{http: context}
+    end
+  end
 end