+++ lib/timber/contexts/runtime_context.ex
@@ -1,10 +1,14 @@
 defmodule Timber.Contexts.RuntimeContext do
-  @moduledoc """
-  The Runtime context tracks information about the current runtime, such as the module, file,
-  function, and line number that called the log line.
+  @moduledoc ~S"""
+  **DEPRECATED**
 
-  Note: This is automatically provided by the default `Elixir.Logger`. There is no need
-  to add this yourself.
+  This module is deprecated in favor of using `map`s. The next evolution of Timber (2.0)
+  no long requires a strict schema and therefore simplifies how users set context:
+
+      Timber.add_context(runtime: %{application: "my_app", file: "my_file.ex"})
+
+  Please note that this context is added automatically as part of shipping log data.
+  You should not have to do anything to obtain this context.
   """
 
   @type t :: %__MODULE__{
@@ -26,4 +30,11 @@ defmodule Timber.Contexts.RuntimeContext do
         }
 
   defstruct [:application, :file, :function, :line, :module_name, :vm_pid]
+
+  defimpl Timber.Contextable do
+    def to_context(context) do
+      context = Map.from_struct(context)
+      %{runtime: context}
+    end
+  end
 end