+++ lib/timber/contexts/system_context.ex
@@ -1,9 +1,14 @@
 defmodule Timber.Contexts.SystemContext do
-  @moduledoc """
-  The system context tracks OS level information, such as the process ID and hostname.
+  @moduledoc ~S"""
+  **DEPRECATED**
 
-  Note: This context is automatically added when each `Timber.Logentry` is created.
-  There is no need to add it yourself.
+  This module is deprecated in favor of using `map`s. The next evolution of Timber (2.0)
+  no long requires a strict schema and therefore simplifies how users set context:
+
+      Timber.add_context(system: %{hostname: "hostname", pid: "abcd1234"})
+
+  Please note that this context is set automatically for you as part of this library.
+  You should not need to do anything to obtain this context.
   """
 
   @type t :: %__MODULE__{
@@ -17,4 +22,11 @@ defmodule Timber.Contexts.SystemContext do
         }
 
   defstruct [:hostname, :pid]
+
+  defimpl Timber.Contextable do
+    def to_context(context) do
+      context = Map.from_struct(context)
+      %{system: context}
+    end
+  end
 end